requirements.txt
aniso8601==7.0.0
anysearch==0.2.1
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2022.6.15
Django==4.0.7
django-cors-headers==3.13.0
django-encrypted-json-fields==1.0.4
django-environ==0.9.0
django-filter==22.1
django-graphiql==0.4.4
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-ratelimit==3.0.1
elasticsearch==7.17.6
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
iso8601==1.0.2
lxml==4.9.1
pandas==1.4.4
pdfkit==1.0.0
promise==2.3
PyJWT==1.7.1
python-dateutil==2.8.2
python-magic==0.4.27
redis==4.3.4
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
stringcase==1.2.0
text-unidecode==1.3
tzdata==2022.2
urllib3==1.26.12
requests~=2.28.1
Pillow==9.2.0
genson==1.2.2
xmltodict==0.13.0
dicttoxml==1.7.15
django-storages==1.13.2
boto3==1.26.76
instamojo_wrapper==1.2.0
